+++ src/components/editors/EditorArea.js
@@ -104,7 +104,7 @@ const EditorArea = ({
 
         <div className="edit-area" /*style={{width: "49%", display: "inline-block"}}*/>
 
-            <Button variant="primary" onClick={e => handleConvert(e)}>Convert</Button>
+            <Button style={{margin: "15px 3px"}} variant="primary" onClick={e => handleConvert(e)}>Convert</Button>
 
             <div className="small-editor-div"/*className="node-edge-editor"*/>
                 <h5>Global defaults editor</h5>